# docs/openapi/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "job-tracker-api",
    "description": "REST API for job-tracker-dashboard",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me/analysis-plan": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get current user's analysis plan",
        "description": "Return current analysis plan limits and feature flags.\n\nThis is read-only. Plan changes are intentionally not exposed as public\nself-upgrade until billing/admin authorization is implemented.",
        "operationId": "get_my_analysis_plan_api_v1_users_me_analysis_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPlanRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/dev/users/me/analysis-plan": {
      "patch": {
        "tags": [
          "dev-tools"
        ],
        "summary": "Development-only current user analysis plan switch",
        "description": "Set current user's analysis plan for local development testing only.",
        "operationId": "patch_my_analysis_plan_for_dev_api_v1_dev_users_me_analysis_plan_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevAnalysisPlanPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevAnalysisPlanUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/vacancy-import/preview": {
      "post": {
        "tags": [
          "vacancy-import"
        ],
        "summary": "Preview a single user-provided vacancy URL",
        "description": "Fetch and parse one user-provided vacancy URL for an editable preview.\n\nThe endpoint does not create or persist applications and does not store imported HTML.",
        "operationId": "preview_vacancy_import_api_v1_vacancy_import_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyImportPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyImportPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/discovery-preview": {
      "post": {
        "tags": [
          "discovery-preview"
        ],
        "summary": "Preview a manual source input without saving it",
        "operationId": "preview_discovery_source_api_v1_discovery_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/applications/{app_id}/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "List application history",
        "operationId": "list_history_api_v1_applications__app_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "APPLICATION_CREATED",
                    "STATUS_CHANGE",
                    "FIELD_CHANGE",
                    "COMMENT",
                    "APPLICATION_ARCHIVED",
                    "DOCUMENT_UPLOADED",
                    "DOCUMENT_DELETED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by history event type",
              "title": "Type"
            },
            "description": "Filter by history event type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{app_id}/comments": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Add a comment to an application",
        "operationId": "add_comment_api_v1_applications__app_id__comments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisPlanFeatures": {
        "properties": {
          "cover_letter": {
            "type": "string",
            "enum": [
              "disabled",
              "short_template",
              "enabled"
            ],
            "title": "Cover Letter"
          },
          "interview_questions": {
            "type": "boolean",
            "title": "Interview Questions"
          },
          "cv_keywords": {
            "type": "boolean",
            "title": "Cv Keywords"
          },
          "multi_match_comparison": {
            "type": "boolean",
            "title": "Multi Match Comparison"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "high"
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "cover_letter",
          "interview_questions",
          "cv_keywords",
          "multi_match_comparison",
          "priority"
        ],
        "title": "AnalysisPlanFeatures"
      },
      "AnalysisPlanLimits": {
        "properties": {
          "basic_daily_limit": {
            "type": "integer",
            "title": "Basic Daily Limit"
          },
          "ai_daily_limit": {
            "type": "integer",
            "title": "Ai Daily Limit"
          }
        },
        "type": "object",
        "required": [
          "basic_daily_limit",
          "ai_daily_limit"
        ],
        "title": "AnalysisPlanLimits"
      },
      "AnalysisPlanRead": {
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "basic",
              "premium"
            ],
            "title": "Plan"
          },
          "limits": {
            "$ref": "#/components/schemas/AnalysisPlanLimits"
          },
          "features": {
            "$ref": "#/components/schemas/AnalysisPlanFeatures"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "limits",
          "features"
        ],
        "title": "AnalysisPlanRead"
      },
...
      "CommentRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "feedback_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Type"
          },
          "sentiment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment"
          },
          "rejection_reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason Code"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CommentRequest"
      },
...
      "DevAnalysisPlanPatch": {
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "basic",
              "premium"
            ],
            "title": "Plan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan"
        ],
        "title": "DevAnalysisPlanPatch"
      },
      "DevAnalysisPlanUpdateResponse": {
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "basic",
              "premium"
            ],
            "title": "Plan"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "message"
        ],
        "title": "DevAnalysisPlanUpdateResponse"
      },
      "DiscoveryPreviewMatchPayload": {
        "properties": {
          "loop_id": {
            "type": "string",
            "title": "Loop Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "vacancy_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vacancy Description"
          },
          "confidence": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Confidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "status": {
            "type": "string",
            "const": "saved",
            "title": "Status",
            "default": "saved"
          }
        },
        "type": "object",
        "required": [
          "loop_id",
          "source_id",
          "source_url",
          "source",
          "company_name",
          "role_title",
          "location_text",
          "vacancy_description",
          "confidence",
          "warnings"
        ],
        "title": "DiscoveryPreviewMatchPayload"
      },
      "DiscoveryPreviewRequest": {
        "properties": {
          "loop_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Loop Id"
          },
          "source_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Source Id"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "loop_id",
          "source_id",
          "url"
        ],
        "title": "DiscoveryPreviewRequest"
      },
      "DiscoveryPreviewResponse": {
        "properties": {
          "loop_id": {
            "type": "string",
            "title": "Loop Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "limited"
            ],
            "title": "Status"
          },
          "normalized_url": {
            "type": "string",
            "title": "Normalized Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "can_create_match": {
            "type": "boolean",
            "title": "Can Create Match"
          },
          "match": {
            "$ref": "#/components/schemas/DiscoveryPreviewMatchPayload"
          }
        },
        "type": "object",
        "required": [
          "loop_id",
          "source_id",
          "status",
          "normalized_url",
          "title",
          "company",
          "location",
          "snippet",
          "warnings",
          "can_create_match",
          "match"
        ],
        "title": "DiscoveryPreviewResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItemRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "from_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Status"
          },
          "to_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Status"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "old_value": {
            "title": "Old Value"
          },
          "new_value": {
            "title": "New Value"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "feedback_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Type"
          },
          "sentiment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment"
          },
          "rejection_reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason Code"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "user_id",
          "actor",
          "type",
          "from_status",
          "to_status",
          "field_path",
          "old_value",
          "new_value",
          "comment",
          "feedback_type",
          "sentiment",
          "rejection_reason_code",
          "correlation_id",
          "created_at"
        ],
        "title": "HistoryItemRead"
      },
      "HistoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "HistoryListResponse"
      },
...
      "VacancyImportPreviewRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "VacancyImportPreviewRequest"
      },
      "VacancyImportPreviewResponse": {
        "properties": {
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "vacancy_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vacancy Description"
          },
          "confidence": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Confidence"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source_url",
          "source",
          "company_name",
          "role_title",
          "location_text",
          "vacancy_description",
          "confidence",
          "warnings"
        ],
        "title": "VacancyImportPreviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}